Add Changes 07.01.22
Add "TIMER" Text to the Home Screen (need to finish the Home Screen and to add a Menu)
Create home_screen.cpp/.h and move all Functions for the Home Screen there
Create end_screen.cpp/.h and move all Functions for the End Screen there
Create print_timer.cpp/.h and move all Functions for the Printing of the Timer there
