+++ ASCII_TIMER/ASCII_TIMER_VS/ASCII_TIMER_VS/ASCII_TIMER_VS.vcxproj.filters
@@ -33,6 +33,15 @@
     <ClCompile Include="print_seconds.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="home_screen.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="end_screen.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="print_timer.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="other_functions.h">
@@ -48,6 +57,15 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="print_seconds.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="home_screen.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="end_screen.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="print_timer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
