+++ ASCII_TIMER/ASCII_TIMER_VS/ASCII_TIMER_VS/ascii_timer.cpp
@@ -1,6 +1,6 @@
-#include "other_functions.h"
+#include "home_screen.h"
 
 int main() {
-    printTimer();
-    return 0;
+	homeScreen();
+	return 0;
 }
