+++ ASCII_TIMER/ASCII_TIMER_VS/ASCII_TIMER_VS/end_screen.h
@@ -1,0 +1,6 @@
+#pragma once
+#ifndef END_SCREEN_H
+#define END_SCREEN_H
+void printEndScreen(bool invalidEntry);
+void endScreen(bool invalidEntry);
+#endif
