+++ ASCII_TIMER/ASCII_TIMER_VS/ASCII_TIMER_VS/home_screen.h
@@ -1,0 +1,11 @@
+#pragma once
+#ifndef HOME_SCREEN_H
+#define HOME_SCREEN_H
+void homeScreen();
+void printHomeScreenAndMenu();
+void printT(int line);
+void printI();
+void printM(int line);
+void printE(int line);
+void printR(int line);
+#endif
