+++ ASCII_TIMER/ASCII_TIMER_VS/ASCII_TIMER_VS/other_functions.h
@@ -4,11 +4,4 @@
 void screenClear();
 int tensOfNumberCount(int n);
 int* transformHoursMinutesSeconds(int n);
-void printTimer(int hours, int minutes, int seconds, int n);
-void printTimer();
-void printEndMessage(bool invalidEntry);
-void whenTimerIsZero(bool invalidEntry);
-void lessThan10Sec(int seconds, int n);
-void printOnlySeconds(int seconds);
-void printIt(int hours, int minutes, int seconds, int n);
 #endif
