+++ ASCII_TIMER/ASCII_TIMER_VS/ASCII_TIMER_VS/print_timer.h
@@ -1,0 +1,8 @@
+#pragma once
+#ifndef PRINT_TIMER_H
+#define PRINT_TIMER_H
+void printTimer(int hours, int minutes, int seconds, int n);
+void lessThan10Sec(int seconds, int n);
+void printOnlySeconds(int seconds);
+void printIt(int hours, int minutes, int seconds, int n);
+#endif
